chore: Set up database in circleci
We faced a challenge with using a python orb as well as a database.
Turns out that we could not reach the database from the python
container.
As it is now, we are able to access both as we reverted to using docker
images for both python and postgres, by introducing an executor.